+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 
 - add post categories table migration
 - add post categories rls policies migration
+- add posts categories table migration
+- add posts categories rls policies migration
 
 ### Changed